+++ tests/testthat/test-EvidenceSynthesisModule.R
@@ -406,8 +406,10 @@ test_that("Check prediction intervals", {
       SqlRender::snakeCaseToCamelCaseNames()
     data <- data |>
       filter(!is.na(pi95Lb) & !is.na(pi95Ub)) |>
-      mutate(seLogPi = (log(pi95Ub) - log(pi95Lb)) / 2 * qnorm(0.975),
-             seLogCalPi = (log(calibratedPi95Ub) - log(calibratedPi95Lb)) / 2 * qnorm(0.975))
+      mutate(
+        seLogPi = (log(pi95Ub) - log(pi95Lb)) / 2 * qnorm(0.975),
+        seLogCalPi = (log(calibratedPi95Ub) - log(calibratedPi95Lb)) / 2 * qnorm(0.975)
+      )
 
     expect_true(nrow(data) > 0)
 
@@ -416,7 +418,9 @@ test_that("Check prediction intervals", {
     expect_true(all(data$ci95Ub <= data$pi95Ub))
 
     # Calibrated PI should be wider than uncalibrated PI (but may be shifted, so using SE):
-    expect_true(all(data$seLogCalPi >= data$seLogPi))
+    expect_true(all(data$seLogCalPi >= data$seLogPi, na.rm = TRUE))
+    # (Test only works if most are not NA:)
+    expect_true(mean(is.na(data$seLogCalPi)) < 0.1 & mean(is.na(data$seLogPi)) < 0.1)
   }
 })