+++ NEWS.md
@@ -1,3 +1,21 @@
+Strategus 1.5.0
+===============
+
+New Features
+
+- Update CohortMethod module to support CohortMethod v6 (#277)
+- Update Characterization module to support Characterization v3 (#288)
+- Add prediction intervals to EvidenceSynthesisModule (#276)
+- Add covariate balance meta-analysis to EvidenceSynthesisModule (#275)
+- Add analysisId and stratify options to TreatmentPatterns (#271) 
+- Add support for cohort templates (#242)
+- createResultDataModel should always create the databaseMetadata table (#232)
+
+Fixes
+
+- Uploading fails when results.zip is missing for CM/SCCS (#263)
+
+
 Strategus 1.4.1
 ===============
 
@@ -7,7 +25,7 @@ Strategus 1.4.0
 ===============
 
 - Update TreatmentPatterns module to support >= v3.1.0 (#233)
-- Update SelfConrolledCaseSeries module to support >= v6.0.0 (#231)
+- Update SelfControlledCaseSeries module to support >= v6.0.0 (#231)
 - Update PatientLevelPrediction module to support >= v6.5.0 and to skip diagnostics (#247)
 - Adds more flexible analyses options for Characterization module (#248)
 - Fix spelling of `createEmptyAnalysisSpecifications` function (#249)